Move chain of trust functions into their own module
